pyproject.toml
[project]
name = "hu-tao-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.7.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-dotenv>=1.2.1",
    "watchfiles>=1.1.1",
]
